Refactored db and rpc api
