+++ orchestrator/db/kv/verified_slot_info.go
@@ -1,0 +1,48 @@
+package kv
+
+import (
+	"github.com/boltdb/bolt"
+	"github.com/lukso-network/lukso-orchestrator/shared/bytesutil"
+	"github.com/lukso-network/lukso-orchestrator/shared/types"
+)
+
+// VerifiedSlotInfo
+func (s *Store) VerifiedSlotInfo(slot uint64) (*types.SlotInfo, error) {
+	if v, ok := s.verifiedSlotInfoCache.Get(slot); v != nil && ok {
+		return v.(*types.SlotInfo), nil
+	}
+	var slotInfo *types.SlotInfo
+	err := s.db.View(func(tx *bolt.Tx) error {
+		bkt := tx.Bucket(verifiedSlotInfosBucket)
+		key := bytesutil.Uint64ToBytesBigEndian(slot)
+		value := bkt.Get(key[:])
+		if value == nil {
+			return nil
+		}
+		return decode(value, &slotInfo)
+	})
+	return slotInfo, err
+}
+
+// SaveVerifiedSlotInfo
+func (s *Store) SaveVerifiedSlotInfo(slot uint64, slotInfo *types.SlotInfo) error {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	// storing consensus info into cache and db
+	return s.db.Update(func(tx *bolt.Tx) error {
+		bkt := tx.Bucket(verifiedSlotInfosBucket)
+		slotBytes := bytesutil.Uint64ToBytesBigEndian(slot)
+		enc, err := encode(slotInfo)
+		if err != nil {
+			return err
+		}
+		if status := s.verifiedSlotInfoCache.Set(slot, slotInfo, 0); !status {
+			log.WithField("slot", slot).Warn("could not store verified slot info into cache")
+		}
+		if err := bkt.Put(slotBytes, enc); err != nil {
+			return err
+		}
+		return nil
+	})
+}
