+++ orchestrator/vanguardchain/handler_test.go
@@ -34,14 +34,14 @@ func TestService_OnNewPendingVanguardBlock(t *testing.T) {
 			ProposerSlashings: []*eth.ProposerSlashing{},
 			VoluntaryExits:    []*eth.SignedVoluntaryExit{},
 			PandoraShard: []*eth.PandoraShard{{
-				ParentHash: make([]byte, 32),
-				TxHash: make([]byte, 32),
-				StateRoot: make([]byte, 32),
+				ParentHash:  make([]byte, 32),
+				TxHash:      make([]byte, 32),
+				StateRoot:   make([]byte, 32),
 				BlockNumber: slot,
 				ReceiptHash: make([]byte, 32),
-				Signature: make([]byte, 96),
-				Hash: make([]byte, 32),
-				SealHash: make([]byte, 32),
+				Signature:   make([]byte, 96),
+				Hash:        make([]byte, 32),
+				SealHash:    make([]byte, 32),
 			}},
 		},
 	}
